User authentication, testing, and everything

+++ app.py
@@ -6,7 +6,9 @@ import simplejson
 from datetime import datetime
 
 import common
+# Add ConPaaS src to PYTHONPATH
 common.extend_path()
+import auth
 import actions
 
 app = Flask(__name__)
@@ -14,45 +16,113 @@ app.config['SQLALCHEMY_DATABASE_URI'] = common.config.get(
     'director', 'DATABASE_URI')
 db = SQLAlchemy(app)
 
-@app.route("/start/<servicetype>")
+def __user_auth():
+    """Return a User object on proper user authentication. False otherwise."""
+    user = auth.auth_user(request.values.get('username', ''), 
+        request.values.get('password', ''))
+
+    if user:
+        return user
+
+    return False
+
+@app.route("/start/<servicetype>", methods=['POST'])
 def start(servicetype):
-    s = Service(name="New %s service" % servicetype,
-                type=servicetype)
+    """eg: POST /start/php
+
+    POSTed values must contain username and password.
+
+    Returns a dictionary with service data (manager's vmid and IP address,
+    service name and ID) in case of successful authentication and correct
+    service creation. An empty dictionary is returned otherwise.
+    """
+    user = __user_auth()
+
+    if not user:
+        # Authentication failed
+        return jsonify({})
+
+    # New service with default name, proper servicetype and user relationship
+    s = Service(name="New %s service" % servicetype, type=servicetype, 
+        user=user)
                 
-    s.manager, s.vmid = actions.start(servicetype, s.sid)
     db.session.add(s)
+    # flush() is needed to get auto-incremented sid
+    db.session.flush()
+    s.manager, s.vmid = actions.start(servicetype, s.sid)
     db.session.commit()
     return jsonify(s.to_dict())
 
-@app.route("/stop/<int:serviceid>")
+@app.route("/stop/<int:serviceid>", methods=['POST'])
 def stop(serviceid):
-    s = Service.query.filter_by(sid=serviceid).first()
-    if s:
-        actions.stop(s.vmid)
-        db.session.delete(s)
-        db.session.commit()
-        return simplejson.dumps(True)
+    """eg: POST /stop/3
+
+    POSTed values must contain username and password.
+
+    Returns a boolean value. True in case of successful authentication and
+    proper service termination. False otherwise.
+    """
+    user = __user_auth()
+
+    if user:
+        # Authentication succeeded
+        s = Service.query.filter_by(sid=serviceid).first()
+        if s and s in user.services:
+            # If a service with id 'serviceid' exists and user is the owner
+            actions.stop(s.vmid)
+            db.session.delete(s)
+            db.session.commit()
+            return simplejson.dumps(True)
 
     return simplejson.dumps(False)
 
 @app.route("/download/ConPaaS.tar.gz", methods=['GET'])
 def download():
+    """GET /download/ConPaaS.tar.gz
+
+    Returns ConPaaS tarball.
+    """
     return helpers.send_from_directory(os.path.dirname(__file__), 
         "ConPaaS.tar.gz")
 
 @app.route("/callback/decrementUserCredit.php", methods=['POST'])
 def credit():
-    service_id, decrement = request.values['sid'], request.values['decrement']
-    return jsonify({ 'error': False })
+    """POST /callback/decrementUserCredit.php
+
+    POSTed values must contain sid and decrement.
+
+    Returns a dictionary with the 'error' attribute set to False if the user
+    had enough credit, True otherwise.
+    """
+    service_id = int(request.values.get('sid', -1))
+    decrement  = int(request.values.get('decrement', 0))
+
+    s = Service.query.filter_by(sid=service_id).first()
+    if not s:
+        # The given service does not exist
+        return jsonify({ 'error': True })
+    
+    # Decrement user's credit
+    s.user.credit -= decrement
+
+    if s.user.credit > -1:
+        # User has enough credit
+        db.session.commit()
+        return jsonify({ 'error': False })
+
+    # User does not have enough credit
+    db.session.rollback()
+    return jsonify({ 'error': True })
 
 @app.route("/callback/terminateService.php")
 def terminate():
+    """To be implemented."""
     pass
 
 class User(db.Model):
     uid = db.Column(db.Integer, primary_key=True, 
         autoincrement=True)
-    username = db.Column(db.String(80), unique=True)
+    username = db.Column(db.String(80), unique=True, nullable=False)
     fname = db.Column(db.String(256))
     lname = db.Column(db.String(256))
     email = db.Column(db.String(256), unique=True)
@@ -78,6 +148,9 @@ class Service(db.Model):
     created = db.Column(db.DateTime)
     manager = db.Column(db.String(512))
     vmid = db.Column(db.String(256))
+
+    user_id = db.Column(db.Integer, db.ForeignKey('user.uid'))
+    user = db.relationship('User', backref=db.backref('services', lazy="dynamic"))
 
     def __init__(self, **kwargs):
         # Default values
